pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "oommfc"
version = "0.64.1"
description = "OOMMF calculator."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}

authors = [
    {name = "Marijan Beg"},
    {name = "Martin Lang"},
    {name = "Ryan A. Pepper"},
    {name = "Thomas Kluyver"},
    {name = "Samuel Holt"},
    {name = "Swapneel Amit Pathak"},
    {name = "Hans Fangohr"}
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Education",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Visualization"
]

dependencies = [
    "micromagneticmodel>=0.63.2",
    "micromagnetictests>=0.63.1",
    "ubermagtable>=0.61.3",
    "ubermagutil>=0.63.2",
    "micromagneticdata>=0.65.1"
]

[project.optional-dependencies]
dev = [
    "build",
    "invoke",
    "nbval",
    "pre-commit",
    "pytest-cov",
    "twine"
]

[project.urls]
homepage = "https://ubermag.github.io"
documentation = "https://ubermag.github.io/documentation/oommfc"
repository = "https://github.com/ubermag/oommfc"




[tool.coverage.run]
omit = ["oommfc/tests/*"]

[tool.ruff.lint]
ignore-init-module-imports = true  # do not remove unused imports in __init__ and warn instead
select = [
    "B",   # flake8-bugbear
    "E",   # pycodestyle
    "F",   # Pyflakes
    "I",   # isort
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
]
ignore = [
    # conflict with other rules
    "D203",  # one-blank-line-before-class (conflicts with D204)
    "D212",  # multi-line-summary-first-line (conflicts with D213)
    # conflict with formatter
    "D206",  # indent-with-spaces
    "D300",  # triple-single-quotes
    "E111",  # indentation-with-invalid-multiple
    "E114",  # indentation-with-invalid-multiple-comment
    "E117",  # over-indented
    # conflict with Python 3.6 compatibility
    "UP022", # replace-stdout-stderr
]

[tool.ruff.lint.isort]
known-local-folder = ["oommfc"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = [
    "B018",  # "Found useless expression. Either assign it to a variable or remove it."; false positives when using implicit __repr__ in the notebook
    "E501",  # line too long
    "F811",  # 'redefined-while-unused'; many false positives in notebooks because ipywidgets decorated functions are not recognised
]

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore:((.|\n)*)Sentinel is not a public part of the traitlets API((.|\n)*)",  # dependency of k3d
]

[tool.setuptools.packages.find]
include = ["oommfc*"]

# to define additional information we use a custom table
# see: https://www.python.org/dev/peps/pep-0518/#tool-table
[tool.ubermag]
doi = "10.5281/zenodo.3539461"
copyright_holder = "Marijan Beg and Hans Fangohr"
contributors = [
    "Mark Vousden",
    "Sergii Mamedov",
    "Rebecca Carey",
    "Marc-Antonio Bisotti",
    "Vanessa Nehruji"
]
about = """
- An Object Oriented MicroMagnetic Framework [OOMMF](https://math.nist.gov/oommf/) calculator for computational magnetism models defined with `micromagneticmodel`.
"""
